Fix issue in CMC implementation on systems with signed chars.

+++ crypto/cmc.hh
@@ -26,8 +26,8 @@ cmc_encrypt(const BlockCipher *c,
     uint8_t m[BlockCipher::blocksize];
     uint8_t carry = 0;
     for (size_t j = BlockCipher::blocksize; j != 0; j--) {
-        uint16_t a = (*ctext)[j - 1] ^
-                     (*ctext)[j - 1 + ptext.size() - BlockCipher::blocksize];
+        uint16_t a = (uint8_t)(*ctext)[j - 1] ^
+                     (uint8_t)(*ctext)[j - 1 + ptext.size() - BlockCipher::blocksize];
         m[j - 1] = carry | (uint8_t) (a << 1);
         carry = a >> 7;
     }
@@ -71,8 +71,8 @@ cmc_decrypt(const BlockCipher *c,
     uint8_t m[BlockCipher::blocksize];
     uint8_t carry = 0;
     for (size_t j = BlockCipher::blocksize; j != 0; j--) {
-        uint16_t a = (*ptext)[j - 1] ^
-                     (*ptext)[j - 1 + ctext.size() - BlockCipher::blocksize];
+        uint16_t a = (uint8_t)(*ptext)[j - 1] ^
+                     (uint8_t)(*ptext)[j - 1 + ctext.size() - BlockCipher::blocksize];
         m[j - 1] = carry | (uint8_t) (a << 1);
         carry = a >> 7;
     }
